+++ src/challenge/find_modes/sorting.py
@@ -4,7 +4,7 @@ import numpy as np
 from numpy.typing import NDArray
 
 
-def find_modes_via_sorting(xs: NDArray[np.int32]) -> list[int]:
+def find_modes_via_sorting(xs: NDArray[np.int16]) -> list[int]:
     xs.sort()
 
     max_n = 0
@@ -14,7 +14,7 @@ def find_modes_via_sorting(xs: NDArray[np.int32]) -> list[int]:
     return [value for n, value in get_runs(xs) if max_n == n]
 
 
-def get_runs(xs: NDArray[np.int32]) -> Generator[tuple[int, int]]:
+def get_runs(xs: NDArray[np.int16]) -> Generator[tuple[int, int]]:
     """Generates a (count, value) tuple for each run of an observed value."""
     n = 0
     prev = xs[0]  # input shall be non-empty
